% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryindex.R
\name{binaryindex}
\alias{binaryindex}
\title{Compute the Binary Scope Scale}
\usage{
binaryindex(x)
}
\arguments{
\item{x}{Output file from previous run - \code{dataprep} function}
}
\value{
In addition to the input columns, a `bin` variable
}
\description{
This function read the output of the \code{dataprep} function
and compute the Binary scope scale.
}
\examples{
binaryindex(x)

}
