% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sev.R
\name{calc_sev}
\alias{calc_sev}
\title{Compute the PHSM Severity Index}
\usage{
calc_sev(x)
}
\arguments{
\item{x}{CSV file with binary and ordinal scales scores}
}
\value{
Computed PHSM severity index
}
\description{
This function takes the binary and ordinal scores
and compute the PHSM severity index. See WHO Euro
}
\examples{
calc_sev(x)

}
