% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{dataprep}
\alias{dataprep}
\title{Prepare Data Set for the Severity Calculation}
\usage{
dataprep(x, country = NULL)
}
\arguments{
\item{x}{csv file - mistress or master file}

\item{country=NULL}{Filter for a specific country, else all.}
}
\value{
8 variables required for the calculation of the SI:
}
\description{
This function read a csv data (PHSM mistress or master files)
and return only the variables needed for the calculation
of the severity index (SI).
}
\examples{
dataprep(x)
dataprep(x,'Algeria')

}
