% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forplot.R
\name{forplot}
\alias{forplot}
\title{Call the function that generate the final data}
\usage{
forplot(x, country = NULL)
}
\description{
Call the function that generate the final data
}
\examples{
forplot(x)
forplot(x,country)
}
