% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordbusiness.R
\name{ordbusiness}
\alias{ordbusiness}
\title{Compute the Ordinal Scope Scale for Businesses}
\usage{
ordbusiness(x)
}
\arguments{
\item{x}{Output file from previous run - \code{ordschools} function}
}
\value{
Computed ordinal scale for Businesses category
}
\description{
This function read the output of the \code{ordschools} function
and compute the Ordinal scope scale for Indicator 4.2.1 and 4.2.2.
}
\examples{
ordbusiness(x)

}
