% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordgather.R
\name{ordgather}
\alias{ordgather}
\title{Compute the Ordinal Scope Scale for Gathering}
\usage{
ordgather(x)
}
\arguments{
\item{x}{Output file from previous run - \code{ordbusiness} function}
}
\value{
Computed ordinal scale for Gathering category
}
\description{
This function read the output of the \code{ordbusiness} function
and compute the Ordinal scope scale for Indicator 4.3.1,4.3.2,4.3.3,4.3.4.
}
\examples{
ordgather(x)

}
