% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordmasks.R
\name{ordmasks}
\alias{ordmasks}
\title{Compute the Ordinal Scope Scale for Masks}
\usage{
ordmasks(x)
}
\arguments{
\item{x}{Output file from previous run - \code{binaryindex} function}
}
\value{
In addition to the input columns, a `ord` variable
}
\description{
This function read the output of the \code{binaryindex} function
and compute the Ordinal scope scale for Indicator 1.4.
}
\examples{
ordmasks(x)

}
