% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordmove.R
\name{ordmove}
\alias{ordmove}
\title{Compute the Ordinal Scope Scale for Movements}
\usage{
ordmove(x)
}
\arguments{
\item{x}{Output file from previous run - \code{ordgather} function}
}
\value{
Computed ordinal scale for Movements category
}
\description{
This function read the output of the \code{ordgather} function
and compute the Ordinal scope scale for Indicator 4.5.1,4.5.2,4.5.3,4.5.4.
}
\examples{
ordmove(x)

}
