% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordschools.R
\name{ordschools}
\alias{ordschools}
\title{Compute the Ordinal Scope Scale for Schools}
\usage{
ordschools(x)
}
\arguments{
\item{x}{Output file from previous run - \code{ordmasks} function}
}
\value{
Computed ordinal scale for Schools category
}
\description{
This function read the output of the \code{ordmasks} function
and compute the Ordinal scope scale for Indicator 4.1.1 and 4.1.2.
}
\examples{
ordschools(x)

}
