% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordtravel.R
\name{ordtravel}
\alias{ordtravel}
\title{Compute the Ordinal Scope Scale for International travel}
\usage{
ordtravel(x)
}
\arguments{
\item{x}{Output file from previous run - \code{ordmove} function}
}
\value{
Computed ordinal scale for Travel category
}
\description{
This function read the output of the \code{ordmove} function
and compute the E_scale, Q_scale and Ordinal scope scale for Indicators 5.2','5.3','5.5','5.7','5.8','5.9'.
}
\examples{
ordtravel(x)

}
