% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_plot.R
\name{prep_plot}
\alias{prep_plot}
\title{Prepares the data in the appropriate format}
\usage{
prep_plot(x, country)
}
\arguments{
\item{x}{CSV file with SI value}
}
\value{
Final csv file
}
\description{
This function reads the SI value and fill it
for all dates, starting from the minimum date in the data
}
\examples{
prep_plot(x)

}
